add: sorting options for decks. update: sorting options is now a helper function.
