+++ docs/.vuepress/styles/download.styl
@@ -1,5 +1,4 @@
 .download-tables
-  margin-top: 20px;
   max-width: 800px;
   position: relative;
   
@@ -9,12 +8,33 @@
     margin: 0;
   }
 
+  .sorting-options {
+    position: sticky;
+    top: 55px;
+    z-index: 9;
+    background-color: white;
+    padding: 30px 0 30px;
+    border-bottom: 1px solid $borderColor;
+
+    .show-sort-row {
+      display: inline-block;
+    }
+
+    select {
+      display: inline;
+      font-size: 16px;
+      margin-left: 10px;
+    }
+  }
+
+
   .download-table {
     border: 1px solid #eaecef;
     box-sizing: border-box;
     width: 100%;
     display: flex;
     flex-wrap: wrap;
+    margin-top: -1px;
 
     &:nth-of-type(odd) {
       background-color: #f6f6f6;
